src/requirements.txt
beautifulsoup4==4.12.3
bleach==6.1.0
bs4==0.0.2
celery==5.4.0
crypto==1.4.1
cryptography==42.0.7
cssselect2==0.7.0
Django==5.0.6
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-debug-toolbar==4.3.0
django-taggit==5.0.1
django-unfold==0.36.0
djangorestframework==3.15.2
dj-database-url==2.2.0
flower==2.0.1
jose==1.0.0
MarkupSafe==2.1.5
oauthlib==3.2.2
pillow==10.3.0
# psycopg2==2.9.9        # windows
psycopg2-binary==2.9.9   # linux
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
setuptools==70.1.0
Werkzeug==3.0.3
# uWSGI==2.0.26